+++ src/Molecules/MultiStepProgress/SubLevel/SubLevel.types.ts
@@ -1,5 +1,4 @@
-import { StepBaseProps } from '../MultiStepProgress.types';
-import { A11yProps } from '../Status/Status.types';
+import { A11yProps, StepBaseProps } from '../shared.types';
 
 export type StepSubLevelProps = StepBaseProps;
 
@@ -7,5 +6,9 @@ export type Props = {
   onStepClick?: (stepName: string) => void;
   steps: StepSubLevelProps[];
 };
+
+export type StepTopLevelProps = {
+  steps?: StepSubLevelProps[];
+} & StepBaseProps;
 
 export type SubLevelComponent = React.FC<Props & A11yProps>;
