+++ src/Molecules/MultiStepProgress/TopLevel/TopLevel.types.ts
@@ -1,24 +1,16 @@
-import { StepSubLevelProps } from '../SubLevel/SubLevel.types';
-import { A11yProps } from '../Status/Status.types';
-import { StepBaseProps } from '../MultiStepProgress.types';
+import { A11yProps } from '../shared.types';
+import { StepTopLevelProps } from '../SubLevel/SubLevel.types';
 
 export type InternalProps = {
   $current?: boolean;
   $isInDrawer?: boolean;
 };
 
-export type StepTopLevelProps = {
-  steps?: StepSubLevelProps[];
-} & StepBaseProps;
-
-type Props = {
+export type Props = {
   drawerOpen?: boolean;
   onStepClick?: (stepName: string) => void;
   onSubStepClick?: (stepName: string) => void;
   onMobileStepClick?: () => void;
   steps?: StepTopLevelProps[];
   isInDrawer?: boolean;
-};
-
-export type ProgressLevelsComponent = React.FC<Props & A11yProps>;
-export type TopLevelComponent = React.FC<Props & A11yProps>;
+} & A11yProps;
